964/integration: resolve A×C contract seams (96 features / 104 methods), C v27 × F fresh-bootstrap pin (27 migrations), filesystem.rs mutation×search halves, E4 content law; CI error registry §1–33 checked by reading

+++ crates/haider-platform/src/process.rs
@@ -390,7 +390,7 @@ fn close_inherited_descriptor_range(first: std::os::raw::c_int) -> bool {
             libc::SYS_close_range,
             first as libc::c_uint,
             libc::c_uint::MAX,
-            0 as libc::c_uint,
+            0_u32,
         ) == 0
     }
 }
